% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dist_matrix.R
\name{fct_dist_matrix}
\alias{fct_dist_matrix}
\title{Internal Distance Matrix Computation}
\usage{
fct_dist_matrix(coords, verbose = TRUE)
}
\arguments{
\item{coords}{matrix or data frame; euclidean coordinates in separate columns}

\item{verbose}{TRUE or FALSE; print to screen?}
}
\value{
distance matrix
}
\description{
Internal Distance Matrix Computation
}
\examples{
fct_dist_matrix(matrix(rnorm(20), nrow = 10, ncol = 2), verbose = TRUE)
}
