% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_exp_uv.R
\name{fct_exp_uv}
\alias{fct_exp_uv}
\title{Internal uv multiplication function}
\usage{
fct_exp_uv(u, v, cores)
}
\arguments{
\item{u}{matrix; u matrix}

\item{v}{matrix; v matrix}

\item{cores}{integer; number of cores to use}
}
\value{
matrix;
}
\description{
Internal uv multiplication function
}
