% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_norm.R
\name{fct_norm}
\alias{fct_norm}
\title{Internal normalization function}
\usage{
fct_norm(comp, eta, ident, cores)
}
\arguments{
\item{comp}{matrix; component to be regularized}

\item{eta}{numeric; column normal penalization coefficient}

\item{ident}{matrix; identity matrix r by r}

\item{cores}{integer; number of cores to use}
}
\value{
normalization
}
\description{
Internal normalization function
}
