% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_objective.R
\name{fct_objective}
\alias{fct_objective}
\title{Internal Objective Function}
\usage{
fct_objective(x, u, v, w, j, lambda, cores)
}
\arguments{
\item{x}{matrix; data matrix}

\item{w}{matrix; distance matrix. If null, computed on the fly.}

\item{lambda}{numeric; penalization parameter}

\item{cores}{integer; number of cores to use}
}
\value{
numeric; objective value
}
\description{
Internal Objective Function
}
