% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opt_adadelta.R
\name{fct_opt_adadelta}
\alias{fct_opt_adadelta}
\title{Internal Adadelta Function}
\usage{
fct_opt_adadelta(gradients, state)
}
\arguments{
\item{gradients}{matrix; gradients}

\item{state}{object; state information}
}
\value{
updates
}
\description{
Internal Adadelta Function
}
