% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opt_adam.R
\name{fct_opt_adam}
\alias{fct_opt_adam}
\title{Internal Adam Function}
\usage{
fct_opt_adam(gradients, state)
}
\arguments{
\item{gradients}{matrix; gradients}

\item{state}{object; state information}
}
\value{
updates
}
\description{
Internal Adam Function
}
\examples{
fct_opt_adam(matrix(rnorm(20), nrow = 10, ncol = 2), state = NULL)
}
