% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opt_adamax.R
\name{fct_opt_adamax}
\alias{fct_opt_adamax}
\title{Internal adamax Function}
\usage{
fct_opt_adamax(gradients, state)
}
\arguments{
\item{gradients}{matrix; gradients}

\item{state}{object; state information}
}
\value{
updates
}
\description{
Internal adamax Function
}
\examples{
fct_opt_adamax(matrix(rnorm(20), nrow = 10, ncol = 2), state = NULL)
}
