% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opt_amsgrad.R
\name{fct_opt_amsgrad}
\alias{fct_opt_amsgrad}
\title{Internal AMSGrad Function}
\usage{
fct_opt_amsgrad(gradients, state)
}
\arguments{
\item{gradients}{matrix; gradients}

\item{state}{object; state information}
}
\value{
updates
}
\description{
Internal AMSGrad Function
}
\examples{
fct_opt_amsgrad(matrix(rnorm(20), nrow = 10, ncol = 2), state = NULL)
}
