% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opt_grad_desc.R
\name{fct_opt_grad_desc}
\alias{fct_opt_grad_desc}
\title{Internal Gradient Descent Function}
\usage{
fct_opt_grad_desc(gradients, state)
}
\arguments{
\item{gradients}{matrix; gradients}

\item{state}{object; state information}
}
\value{
updates
}
\description{
Internal Gradient Descent Function
}
\examples{
fct_opt_grad_desc(matrix(rnorm(20), nrow = 10, ncol = 2), state = NULL)
}
