% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opt_nadam.R
\name{fct_opt_nadam}
\alias{fct_opt_nadam}
\title{Internal Nadam Function}
\usage{
fct_opt_nadam(gradients, state)
}
\arguments{
\item{gradients}{matrix; gradients}

\item{state}{object; state information}
}
\value{
updates
}
\description{
Internal Nadam Function
}
\examples{
fct_opt_nadam(matrix(rnorm(20), nrow = 10, ncol = 2), state = NULL)
}
