% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opt_rmsprop.R
\name{fct_opt_rmsprop}
\alias{fct_opt_rmsprop}
\title{Internal RMSprop Function}
\usage{
fct_opt_rmsprop(gradients, state)
}
\arguments{
\item{gradients}{matrix; gradients}

\item{state}{object; state information}
}
\value{
updates
}
\description{
Internal RMSprop Function
}
\examples{
fct_opt_rmsprop(matrix(rnorm(20), nrow = 10, ncol = 2), state = NULL)
}
