% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optimize.R
\name{fct_optimize}
\alias{fct_optimize}
\title{Internal Optimization Function}
\usage{
fct_optimize(
  x,
  u,
  v,
  w,
  lambda,
  optimizer = "fct_amsgrad",
  epsilon = 1e-08,
  max_iter = 1000,
  verbose = TRUE
)
}
\arguments{
\item{x}{data matrix}

\item{u}{initial u matrix}

\item{v}{initial v matrix}

\item{w}{distance matrix}

\item{lambda}{numeric; penalization parameter}

\item{optimizer}{internal function; only fct_amsgrad for now}

\item{epsilon}{numeric; convergence criterion}

\item{max_iter}{integer; maximum number of iterations}

\item{verbose}{TRUE or FALSE; print to screen?}
}
\value{
list of results
}
\description{
Internal Optimization Function
}
\examples{
fct_optimize(x = matrix(rnorm(100), nrow = 10, ncol = 10), 
u = matrix(rnorm(20), nrow = 10, ncol = 2), 
v = matrix(rnorm(20), nrow = 10, ncol = 2), 
w = matrix(rnorm(100), nrow = 10, ncol = 10), 
lambda = 1, optimizer = "fct_amsgrad", epsilon = 1e-8, max_iter = 10, verbose = TRUE)
}
