% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_push_me.R
\name{fct_push_me}
\alias{fct_push_me}
\title{Pushover Capability for Large Simulation}
\usage{
fct_push_me(message = "complete")
}
\arguments{
\item{message}{a message about script progress}
}
\value{
nothing
}
\description{
Pushover Capability for Large Simulation
}
\examples{
fct_push_me("test")
}
