% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reg.R
\name{fct_reg}
\alias{fct_reg}
\title{Internal regularization function}
\usage{
fct_reg(comp, eta, core)
}
\arguments{
\item{comp}{matrix; component to be regularized}

\item{eta}{numeric; column normal penalization coefficient}
}
\value{
normalization
}
\description{
Internal regularization function
}
