% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_u_grad.R
\name{fct_u_grad}
\alias{fct_u_grad}
\title{Internal U Gradient Function}
\usage{
fct_u_grad(x, u, v, u_penal, uv_exp, w, j2, one, lambda, cores)
}
\arguments{
\item{x}{matrix; data matrix}

\item{u}{matrix; u matrix}

\item{v}{matrix; v matrix}

\item{u_penal}{numeric; penalization if u is normalized}

\item{uv_exp}{matrix; stored matrix multiplication}

\item{w}{matrix; distance matrix. If null, computed on the fly.}

\item{j2}{matrix; matrix of 1s with dimension r by n}

\item{one}{matrix; matrix of 1s with same shape as u}

\item{lambda}{numeric; penalization parameter. set to NULL for lambda selection.}

\item{cores}{integer; number of cores to use}
}
\value{
matrix; u gradient
}
\description{
Internal U Gradient Function
}
