% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_u_grad_c.R
\name{fct_u_grad_c}
\alias{fct_u_grad_c}
\title{Title}
\usage{
fct_u_grad_c(x, u, v, w, j, lambda, cores)
}
\arguments{
\item{x}{t}

\item{u}{t}

\item{v}{t}

\item{w}{t}

\item{j}{t}

\item{lambda}{t}

\item{cores}{t}
}
\value{
t
}
\description{
Title
}
