% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_v_grad.R
\name{fct_v_grad}
\alias{fct_v_grad}
\title{Internal V Gradient Function}
\usage{
fct_v_grad(x, u, v, uv_exp, cores)
}
\arguments{
\item{x}{matrix; data matrix}

\item{uv_exp}{matrix; stored matrix multiplication}

\item{cores}{integer; number of cores to use}
}
\value{
matrix; v gradient
}
\description{
Internal V Gradient Function
}
