% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clust.R
\name{spatial_clust}
\alias{spatial_clust}
\title{Main Spatial Clustering Method}
\usage{
spatial_clust(
  x,
  u_init,
  v_init,
  coords,
  lambda,
  optimizer = "fct_amsgrad",
  epsilon = 1e-08,
  max_iter = 1000,
  verbose = TRUE
)
}
\arguments{
\item{x}{matrix; data matrix}

\item{u_init}{matrix; u initialization matrix}

\item{v_init}{matrix; v initialization matrix}

\item{coords}{matrix or data frame; euclidean coordinates in separate columns}

\item{lambda}{numeric; penalization parameter}

\item{optimizer}{internal function; only fct_amsgrad for now}

\item{epsilon}{numeric; convergence criterion}

\item{max_iter}{integer; maximum number of iterations}

\item{verbose}{TRUE or FALSE; print to screen?}
}
\value{
list including u, v and convergence information
}
\description{
Main Spatial Clustering Method
}
\examples{
spatial_clust(x = matrix(rnorm(100), nrow = 10, ncol = 10), 
u_init = matrix(rnorm(20), nrow = 10, ncol = 2), 
v_init = matrix(rnorm(20), nrow = 10, ncol = 2), 
coords = matrix(rnorm(20), nrow = 10, ncol = 2), 
lambda = 1, optimizer = "fct_amsgrad", epsilon = 1e-8, max_iter = 10, verbose = TRUE)
}
