% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lik.R
\name{fct_lik}
\alias{fct_lik}
\title{Compute the likelihood portion of the objective function}
\usage{
fct_lik(x, u, v, cores)
}
\arguments{
\item{x}{matrix; data matrix}

\item{u}{matrix; u matrix}

\item{v}{matrix; v matrix}

\item{cores}{integer; number of cores to use}
}
\value{
likelihood value
}
\description{
Compute the likelihood portion of the objective function
}
