% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optimize.R
\name{fct_optimize}
\alias{fct_optimize}
\title{Internal Optimization Function}
\usage{
fct_optimize(
  x,
  u,
  v,
  w,
  lambda,
  optimizer,
  epsilon,
  max_iter,
  verbose,
  fast,
  cores
)
}
\arguments{
\item{x}{matrix; data matrix}

\item{w}{matrix; distance matrix}

\item{lambda}{numeric; penalization parameter}

\item{optimizer}{internal function; only fct_amsgrad for now}

\item{epsilon}{numeric; convergence criterion}

\item{max_iter}{integer; maximum number of iterations}

\item{verbose}{TRUE or FALSE; print to screen?}

\item{fast}{TRUE or FALSE; use compiled c?}

\item{cores}{integer; number of cores to use}
}
\value{
list of results
}
\description{
Internal Optimization Function
}
