% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optimize.R
\name{fct_optimize}
\alias{fct_optimize}
\title{Internal Optimization Function}
\usage{
fct_optimize(
  x,
  u,
  v,
  w,
  lambda,
  optimizer,
  epsilon,
  max_iter,
  norm_comp,
  verbose,
  fast,
  cores
)
}
\arguments{
\item{x}{matrix; data matrix}

\item{u}{matrix; u matrix}

\item{v}{matrix; v matrix}

\item{w}{matrix; distance matrix. If null, computed on the fly.}

\item{lambda}{numeric; penalization parameter. set to NULL for lambda selection.}

\item{optimizer}{internal optimization function; options include: fct_opt_amsgrad, fct_opt_adadelt, fct_opt_adam, fct_opt_adamax, fct_opt_grad_desc, fct_opt_nadam, fct_opt_rmsprop.}

\item{epsilon}{numeric; convergence criterion}

\item{max_iter}{integer; maximum number of iterations}

\item{norm_comp}{string; "u" to normalize u, "v" to normalize v, "none" to skip component normalization.}

\item{verbose}{TRUE or FALSE; print to screen?}

\item{fast}{TRUE or FALSE; use compiled c?}

\item{cores}{integer; number of cores to use}
}
\value{
list of results
}
\description{
Internal Optimization Function
}
