% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_penal.R
\name{fct_penal}
\alias{fct_penal}
\title{Compute the penalty portion of the objective function}
\usage{
fct_penal(u, w, j, cores)
}
\arguments{
\item{w}{matrix; distance matrix. If null, computed on the fly.}

\item{cores}{integer; number of cores to use}
}
\value{

}
\description{
Compute the penalty portion of the objective function
}
