% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clust.R
\name{spatial_clust}
\alias{spatial_clust}
\title{Main Spatial Clustering Method}
\usage{
spatial_clust(
  x,
  u_init,
  v_init,
  coords,
  lambda,
  w = NULL,
  optimizer = "fct_amsgrad",
  epsilon = 1e-08,
  max_iter = 1000,
  verbose = TRUE,
  fast = TRUE,
  cores = 1
)
}
\arguments{
\item{x}{matrix; data matrix}

\item{u_init}{matrix; u initialization matrix}

\item{v_init}{matrix; v initialization matrix}

\item{coords}{matrix or data frame; euclidean coordinates in separate columns}

\item{lambda}{numeric; penalization parameter}

\item{w}{matrix; distance matrix}

\item{optimizer}{internal function; only fct_amsgrad for now}

\item{epsilon}{numeric; convergence criterion}

\item{max_iter}{integer; maximum number of iterations}

\item{verbose}{TRUE or FALSE; print to screen?}

\item{fast}{TRUE or FALSE; use compiled c?}

\item{cores}{integer; number of cores to use}
}
\value{
list including u, v and convergence information
}
\description{
Main Spatial Clustering Method
}
